requirements.txt
aiogram
python-dotenv
SQLAlchemy
alembic
asyncpg
aiohttp
ujson